Fix #621 - prevent error on upgrade using mysql.

+++ db/upgrade.php
@@ -32,7 +32,7 @@ require_once(dirname(__FILE__) . '/upgradelib.php');
  */
 function xmldb_attendance_upgrade($oldversion=0) {
 
-    global $DB;
+    global $DB, $CFG;
     $dbman = $DB->get_manager(); // Loads ddl manager and xmldb classes.
 
     $result = true;
@@ -730,11 +730,13 @@ function xmldb_attendance_upgrade($oldversion=0) {
         upgrade_mod_savepoint(true, 2022083100, 'attendance');
     }
     if ($oldversion < 2022090900) {
-        $sql = 'DELETE FROM {attendance_log}
-                 WHERE ID NOT IN (SELECT max(id)
-                                    FROM {attendance_log}
-                                GROUP BY sessionid, studentid, statusid)';
-        $DB->execute($sql);
+        if (!empty($CFG->dbfamily) && $CFG->dbfamily == 'postgres') {
+            $sql = 'DELETE FROM {attendance_log}
+            WHERE ID NOT IN (SELECT max(id)
+                               FROM {attendance_log}
+                           GROUP BY sessionid, studentid, statusid)';
+            $DB->execute($sql);
+        }
 
         // Attendance savepoint reached.
         upgrade_mod_savepoint(true, 2022090900, 'attendance');
